refactor: extract CounriesItem
