+++ src/components/CountriesItem/CountriesItem.scss
@@ -1,10 +1,6 @@
 @import '../../styles/colors.scss';
 
 .countries_list {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-
   &--item {
     width: 200px;
     height: 300px;
