+++ src/components/CountriesItem/classNames.js
@@ -1,3 +1,4 @@
-export const COUNTRIES_LIST = 'countries_list';
+import { COUNTRIES_LIST } from '../../containers/CountriesList/classNames';
+
 export const COUNTRIES_ITEM = `${COUNTRIES_LIST}--item`;
 export const COUNTRIES_BLACKOUT = `${COUNTRIES_LIST}--blackout`;
