+++ src/containers/CountriesList/CountriesList.scss
@@ -1,0 +1,7 @@
+@import '../../styles/colors.scss';
+
+.countries_list {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
