+++ src/containers/CountriesList/classNames.js
@@ -1,0 +1,1 @@
+export const COUNTRIES_LIST = 'countries_list';
