+++ src/pages/MainPage.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Header from '../components/Header/Header';
 import Search from '../components/Search/Search';
 import Wrapper from '../components/Wrapper/Wrapper';
-import CountriesList from '../components/CountriesList/CountriesList';
+import CountriesList from '../containers/CountriesList/CountriesList';
 
 const MainPage = () => (
   <React.Fragment>
